requirements.txt
djangorestframework
djangorestframework-simplejwt
django-cors-headers
Node.js
npm
React
